Move ModelState and ModelDimension classes to separate files
... because thousands of line per file are a pain
