+++ matlab/@amimodel/compileAndLinkModel.m
@@ -196,7 +196,7 @@ function [objectStrAmici] = compileAmiciBase(amiciRootPath, objectFolder, object
     cppsrc = {'amici', 'symbolic_functions','spline', ...
         'edata','rdata', 'exception', ...
         'interface_matlab', 'misc', 'simulation_parameters', ...
-        'solver', 'solver_cvodes', 'solver_idas', ...
+        'solver', 'solver_cvodes', 'solver_idas', 'model_state', ...
         'model', 'model_ode', 'model_dae', 'returndata_matlab', ...
         'forwardproblem', 'steadystateproblem', 'backwardproblem', 'newton_solver', ...
         'abstract_model', 'sundials_matrix_wrapper', 'sundials_linsol_wrapper', ...
