+++ swig/amici.i
@@ -62,6 +62,8 @@ wrap_unique_ptr(ExpDataPtr, amici::ExpData)
 // Include before any other header which uses enums defined there
 %include "amici/defines.h"
 
+%include "amici/model_dimensions.h"
+%include "amici/model_state.h"
 %include "amici/simulation_parameters.h"
 
 %include abstract_model.i
